Унификация расположения файлов. Страница таро.
