+++ README.md
@@ -14,14 +14,15 @@ http://185.80.91.29:8000/
 4. Рулетка для выбора фильма
 5. Открытка с будущими фильмами
 6. Архив открыток с сеансами
+7. Теги для сортировки фильмов
 
 Может быть:
-1. Игровой выбор фильмов (рулетка, турнир, квиз по фильмам)
+1. Игровой выбор фильмов (турнир, квиз по фильмам)
 2. Добавить рубрику "Угадай кино"
 3. Список выпитого чая/коктейлей
-4. Саундтреки при наведении на постер
-5. Теги для сортировки фильмов
-6. Текущее состояние бара
+4. Голосование за фильмы
+5. Текущее состояние бара
+6. Фильтры для статистики
 
 Логи приложения:  
 http://185.80.91.29:3000/
