+++ filmoclub/urls.py
@@ -26,7 +26,7 @@ urlpatterns = [
     path("movies/", include("lists.urls")),
     path("", include("postcard.urls")),
     path("tools/", include("tools.urls")),
-    path("statistic/", include("statistic.urls")),
+
     path("features/", include("features.urls")),
 ]
 urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
