+++ static/css/features/statistic.css
