+++ static/js/postcards_archive.js
@@ -1,10 +1,10 @@
-import {Request} from "../utils/request.js";
-import {confirmModalAction} from "../utils/confirm_modal_action.js";
+import {Request} from "./utils/request.js";
+import {confirmModalAction} from "./utils/confirm_modal_action.js";
 
 // Функция удаления открытки
 async function removePostcard(postcardNode, postcardId) {
 
-    const response = await Request.delete({url: '/', body: {id: postcardId}})
+    const response = await Request.delete({url: 'archive/', body: {id: postcardId}})
 
     if (response !== null) {
