--- statistic/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class StatisticConfig(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "statistic"
