+++ templates/add_movie.html
@@ -6,7 +6,7 @@
     <link rel="stylesheet" href="{% static 'css/add_movie.css' %}">
 {% endblock %}
 
-{% block title %}Фильмы{% endblock %}
+{% block title %}Добавление фильма{% endblock %}
 
 {% block center-content %}
