+++ templates/based_layout.html
@@ -72,13 +72,15 @@
         <div class="col-md-3 h-100">
             <div class="container nav-column">
                 <nav class="nav flex-column">
-                    <a class="nav-link active" aria-current="page" href="/">Церемония</a>
-                    <a class="nav-link" href="/movies/">Список</a>
+{#                    <a class="nav-link active" aria-current="page" href="/">Церемония</a>#}
+                    <a class="nav-link" href="/">Церемония</a>
+                    <a class="nav-link" href="/archive/">Открытки</a>
+                    <a class="nav-link" href="/movies/">Фильмы</a>
                     <a class="nav-link" href="/movies/archive/">Архив</a>
                     <a class="nav-link" href="/movies/add/">Добавление</a>
-                    <a class="nav-link" href="/statistic/">Статистика</a>
+                    <a class="nav-link" href="/features/tarot/">Таро</a>
+                    <a class="nav-link" href="/features/statistic/">Статистика</a>
                     <a class="nav-link" href="/features/carousel/">Карусель</a>
-                    <a class="nav-link" href="/features/postcards_archive/">Открытки</a>
                 </nav>
                 <img class="align-self-center mt-auto nav-img" src="{{ random.navigation }}">
             </div>
