+++ templates/features/features_catalog.html
@@ -1,0 +1,4 @@
+{% extends "based_layout.html" %}
+
+{% load static %}
+
