+++ templates/features/statistic.html
@@ -3,9 +3,9 @@
 {% load static %}
 
 {% block style %}
-    <link rel="stylesheet" href="{% static 'css/statistic.css' %}">
+    <link rel="stylesheet" href="{% static 'css/features/statistic.css' %}">
 {% endblock %}
-{% block title %}Фильмы{% endblock %}
+{% block title %}Статистика{% endblock %}
 
 {% block center-content %}
 
@@ -39,10 +39,11 @@
 
     <div id="stats">
         <ul>
-            <li>Всего посмотренно фильмов: <span>{{ statistic.count }} </span></li>
+            <li>Всего посмотрено фильмов: <span>{{ statistic.count }} </span></li>
             <li>Общая продолжительность: <span>{{ statistic.total_duration }}</span></li>
-            <li>Средняя оценка на кинопоиске: <span>{{ statistic.rating_kp }}</span></li>
+            <li>Средняя оценка на кинополке: <span>{{ statistic.rating_kp }}</span></li>
             <li>Средняя оценка на imdb: <span>{{ statistic.rating_imdb }}</span></li>
+            <li>Средняя оценка на кинопоиске: <span>{{ statistic.rating_kp }}</span></li>
         </ul>
     </div>
