+++ templates/features/tarot.html
@@ -1,0 +1,163 @@
+{% extends "based_layout.html" %}
+
+{% load static %}
+
+{% block style %}
+    <style>
+    
+        body {
+            background: black;
+        }
+    
+        .tarot-container {
+            display: flex;
+            gap: 2rem;
+            max-width: 1200px;
+            margin: 0 auto;
+        }
+
+        .deck {
+            width: 233px;
+            height: 350px;
+            perspective: 1000px;
+            cursor: pointer;
+            flex-shrink: 0;
+        }
+
+        .deck-card {
+            width: 100%;
+            height: 100%;
+            {#background: linear-gradient(45deg, #2c1810, #4a2f24);#}
+            border-radius: 10px;
+            box-shadow: 0 0 20px rgba(0, 0, 0, 0.5);
+            background-image: url('/static/img/tarots/back.jpg');
+            background-size: cover;
+            transition: transform 0.3s ease;
+        }
+
+        .deck:hover .deck-card {
+            transform: translateY(-10px);
+        }
+
+        .reading-area {
+            flex-grow: 1;
+            display: flex;
+            flex-wrap: wrap;
+            gap: 1rem;
+            padding: 20px;
+            min-height: 400px;
+            background: rgba(255, 255, 255, 0.05);
+            border-radius: 15px;
+            overflow: auto;
+        }
+
+        .tarot-card {
+            width: 150px;
+            height: 260px;
+            position: relative;
+            transform-style: preserve-3d;
+            transition: transform 0.8s;
+        }
+
+        .tarot-card.flipped {
+            transform: rotateY(180deg);
+        }
+
+        .card-face {
+            position: absolute;
+            width: 100%;
+            height: 100%;
+            backface-visibility: hidden;
+            border-radius: 10px;
+            box-shadow: 0 0 10px rgba(0, 0, 0, 0.3);
+        }
+
+        .card-back {
+            background: linear-gradient(45deg, #2c1810, #4a2f24);
+            background-image: url('/static/img/tarots/back.jpg');
+            background-size: cover;
+        }
+
+        .card-front {
+            transform: rotateY(180deg);
+            background-size: cover;
+            background-position: center;
+        }
+
+        .instructions {
+            text-align: center;
+            margin-bottom: 2rem;
+            color: #ccc;
+        }
+    </style>
+{% endblock %}
+
+{% block title %}Таро{% endblock %}
+
+
+{# Поле с картами #}
+{% block center-content %}
+    <div class="reading-area" id="readingArea"></div>
+{% endblock %}
+
+{# Колода #}
+{% block right-content %}
+    <div class="deck" id="deck">
+        <div class="deck-card"></div>
+    </div>
+{% endblock %}
+
+
+
+{# Подключаем js #}
+{% block script %}
+    <script>
+
+
+        const cards = [
+            'fool', 'magician', 'high-priestess', 'empress', 'emperor',
+            'hierophant', 'lovers', 'chariot', 'strength', 'hermit',
+            'wheel-of-fortune', 'justice', 'hanged-man', 'death', 'temperance',
+            'devil', 'tower', 'star', 'moon', 'sun', 'judgement', 'world'
+        ];
+
+        const deck = document.getElementById('deck');
+        const readingArea = document.getElementById('readingArea');
+        let drawnCards = [];
+
+        deck.addEventListener('click', () => {
+            if (drawnCards.length >= cards.length) {
+                alert('Все карты уже вытянуты!');
+                return;
+            }
+
+            let randomCard;
+            do {
+                randomCard = cards[Math.floor(Math.random() * cards.length)];
+            } while (drawnCards.includes(randomCard));
+
+            drawnCards.push(randomCard);
+
+            const cardElement = document.createElement('div');
+            cardElement.className = 'tarot-card';
+
+            const cardBack = document.createElement('div');
+            cardBack.className = 'card-face card-back';
+
+            const cardFront = document.createElement('div');
+            cardFront.className = 'card-face card-front';
+            cardFront.style.backgroundImage = `url('/static/img/tarots/${randomCard}.png')`;
+
+            cardElement.appendChild(cardBack);
+            cardElement.appendChild(cardFront);
+            readingArea.appendChild(cardElement);
+
+            // Добавляем небольшую задержку перед переворотом для лучшего визуального эффекта
+            setTimeout(() => {
+                cardElement.classList.add('flipped');
+            }, 100);
+        });
+    </script>
+{% endblock %}
+
+
