+++ templates/postcard.html
@@ -7,7 +7,7 @@
     <link rel="stylesheet" href="{% static 'css/elements/postcard_header.css' %}">
 {% endblock %}
 
-{% block title %}Фильмы{% endblock %}
+{% block title %}Церемония{% endblock %}
 
 {% block header-button %}
     {% include "elements/postcard_header.html" %}
