+++ templates/postcards_archive.html
@@ -3,7 +3,7 @@
 {% load static %}
 
 {% block style %}
-    <link rel="stylesheet" href="{% static 'css/features/postcards_archive.css' %}">
+    <link rel="stylesheet" href="{% static 'css/postcards_archive.css' %}">
 {% endblock %}
 
 {% block title %}Архивные открытки{% endblock %}
